Fixed issue when deleting games
- Server did not properly handle fake players when deleting game data
- Implemented do_turn() of dumb doppelkopf bot

+++ server/cgserver/game/__init__.py
@@ -235,10 +235,10 @@ class CGame(object, metaclass=abc.ABCMeta):
         # Delete all bot players and their corresponding threads
         for pid in self.players:
             p = self.cg.server.users_uuid[pid]
+            if p in self.fake_players:
+                continue
             if isinstance(p, cgserver.user.BotUser):
                 # Stop and delete the bot
                 p.bot.stop()
                 p.bot.delete()
                 del self.cg.server.users_uuid[pid]
-
-
